Make macOS use EditablePartialProfile
